pyproject.toml
[tool.poetry]
name = "first-flask"
version = "0.1.0"
description = "first flask experience"
authors = ["Abra19 <evabrarova@gmail.com>"]
readme = "README.md"
repository = "https://github.com/Abra19/first_flask"
license = "MIT"
packages = [
    { include = "first_flask" },
]
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
